Don't create a check with NIL lisp implementaion
Presumable, the problem occur when a check is created
for the source not bound to any quicklisp distribution.

Previously a check was created with "nil" in the lisp_implementation
slot. Because of that, no worker was able to process such a check.

Also, such check prevented creation of new correct checks
because "check already exists" error.

After the fix check will not be created in this case.