Fixed the database deadlock occured in rare cases
It occured when we were not able to update check's state in the database.

Thanks to this post https://www.reddit.com/r/Common_Lisp/comments/1ik77ra/til_right_way_to_handlerbind_unwind_early_if/ for an idea about the reason of the dead-lock. The problem was in two transaction starting in the same thread because of HANDLER-BIND handler.