Enabling the latest ultralisp dist in qlfile
The dist declared is from June '25. And the thinking is that if I
leave it as that, I'll have to declare every dependency of
fukamachi/qlot individually. So in order to not have to do that, I'm
going with this approach, where fukamachi/qlot will pull it's latest
dependencies from ultralisp.