Start on Quicklisp import
Importing the Quicklisp projects without the moderator email.

Currently only importing GitHub imports.

Previously when source.txt for a project would be parsed, it would
return only a plist with :source & :url, which works for most
cases. But there are projects that are distributed via http, https as
tar balls, git releases, git tags... This is why I added :extra
keyword to the plist.