Improve dev REPL experience
The project is useful but it has many spinning parts.

It has documentation, but it's not super obvious where to look and how
to spin it up to start to contribute to it.

I've already made a bunch of contributions before discovering this
super useful documentation (that I wish I saw earlier):

https://ultralisp.github.io/ultralisp/dev/#x-28ULTRALISP-DOCS-2FDEV-3A-3A-40DEV-2040ANTS-DOC-2FLOCATIVES-3ASECTION-29

And this being a Common Lisp project, it should have a good
interactive, REPL story. Since there was some code, that was commented
out in .local.el, I've tried re-using that and also re-using
docker-tramp. But didn't get very far. Instead I've hacked up
`ultralisp-local.el` along with the update to `.dir-locals.el`.

A word or two about `ultralisp-local.el`:

Why did I move/rename `.local.el` to `ultralisp-local.el`?

Because this project needs a good file path translation story and
there is no need to "hide", so to speak, of Emacs code in
`.local.el`. So created `ultralisp-local.el`, that can be clearly seen
by other devs.

Since this project uses Sly, in docker, that is started via
`docker compose up` which exposes port 14005 for the SLY REPL.
Once you connect to it and you try to `C-c C-k`
(sly-compile-and-load-file), it complains, because the file path
translation between the local file path and that within docker
image, don't match. Tried docker-tramp, didn't work for me. Also,
docker-tramp is additional package. With this approach, you don't need
to load another Emacs package, just to get file path mapping.

And why the need for `sly-ultralisp-enable` & `sly-ultralisp-disable`?
Because `sly-filename-translations` is a global variable, defined in
sly-tramp. So if a dev defines some values in it, it might be useful
to have a way to have values that are not applicable to ultralisp
project kept separate.